Move footer with styles to antoher folder
