+++ src/components/footer/footer.css
@@ -1,0 +1,28 @@
+/*  Styling for Footer Component*/
+.footer{
+    background-color: #f8c829da;
+    margin:0px auto;
+    padding: 20px 0px 20px 0px;
+    color: black;
+  }
+  .footer-links{
+    display:flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 25px;
+  }
+  
+  .footer-links a{
+    color: black;
+    
+  }
+  .footer-btn{
+    margin-top: 15px;
+    margin-bottom: 20px;
+  }
+  
+  .text-center btn{
+    border: 1px solid black;
+    border-radius: 50%;
+  }
+  /*Styling done for footer section*/
